refactor: rename the pipeline façade's nodes module to graph_editor
The module turns a request body into a graph edit; "nodes" said nothing
about that and read as a sibling of the Node model.
